Split up TaskTest into two files for TaskData tests and TaskController tests.
